Abstract export and import queueing
(cherry picked from commit 84745b413550460a45b489078d829462ffe4bcb4)
